Crypto_Agent/crypto_trading_signal_predictor/pyproject.toml
[project]
name = "crypto-trading-signal-predictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ccxt>=4.5.2",
    "chainlit>=2.6.8",
    "fastapi>=0.116.1",
    "matplotlib>=3.10.5",
    "mplfinance>=0.12.10b0",
    "numpy>=2.3.2",
    "openai-agents>=0.2.5",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
]
